feat(textures): Texture.id is a BackendTextureId — #328 phase 3 (bgfx) (#72)
* feat(textures): Texture.id is a BackendTextureId — #328 phase 3 (bgfx)

First backend on the typed id. Pins labelle-core v1.28.0 (the types) and
labelle-gfx v1.29.2.

`Texture.id` is now `core.BackendTextureId`, and `handleForId` /
`nativeTextureHandle` take that type instead of a bare `u32`. Handing
either the engine-facing `TextureId` is now a COMPILE ERROR rather than a
silent invalid-handle return — which is the bug that blanked a
downstream UI kit's menu (labelle-gfx#326).

21 sites in `gfx/texture.zig` converted: slot indices retag explicitly
with `@enumFromInt` when constructing a Texture, and read back through
`.toInt()` when indexing `texture_handles` / `texture_pixel_data`.

Also adds a compile probe to `src/gfx.zig`, which matters more than it
looks. `gfx_tests` roots at that file and Zig only analyses what is
reachable, so `gfx/texture.zig`'s function BODIES were never compiled by
`zig build test` — nor are its inline tests collected. This type change
passed the entire suite while every assignment in that file was still
wrong; three probes inside texture.zig produced no output at all before
I worked out why. The probe forces analysis behind a runtime-false
guard (these calls need a live bgfx context).

Verified the probe bites: passing a `u32` to `nativeTextureHandle` fails
with "expected type 'backend_contract.BackendTextureId', found 'u32'".

* fix: cover the paths CI caught and the probe missed

CI's macOS job failed on `material_golden.zig:296` — a file `zig build
test` never compiles. Fixing it surfaced a second site in
`gfx/render_target.zig` that Zig had not reported, because it stopped
after the first error.

Both are the same untyped surface: `MaterialUniforms.aux_texture` and
`PostPassUniforms.aux_texture` are bare `u32` in the core contract and
carry a BACKEND texture id — `render_target.zig` feeds one straight to
`handleForId`. Retagged at those two boundaries and documented; typing
that surface is follow-up work on labelle-gfx#328, not this PR.

Probe expanded per CodeRabbit's review to call every remaining public
texture path — createDynamicTexture, updateTexture, uploadCompressed,
drawTexturePro, drawTextureProMaterial, drawExternalTexture, the four
YUV plane functions, and the rest. Its point was exactly right: the
narrow probe passed while typed-id errors sat in bodies it never
reached.

All four CI build steps now run locally: test, material-golden,
post-fx-golden, post-fx-integration-golden.

* fix: retag the video sentinel and unify core onto backend_gfx

Both P1s from codex on #72, and both would have broken real consumers
rather than CI.

1. `video/player.zig` built a `types.Texture` with a bare `u32` sentinel.
   `VideoPlayer` is a `pub const` alias to a GENERIC, so neither the alias
   nor the expanded probe analysed `Player.init`'s body — any consumer
   building the in-engine video path would have hit it first. Now `.none`,
   the reserved invalid `BackendTextureId`, and the probe instantiates
   `VideoPlayer(DesktopVideoDecoder)` so the body is compiled here.

2. `build_fragments/backend_dep.txt` overrode `labelle-core` only beneath
   `sdl_gamepad`. In a generated game the app's core and the backend
   dependency's are separate module instances, and Zig treats their enums
   as distinct nominal types even from identical source — so the
   documented flow

       const backend_id = game.nativeTextureId(handle) orelse return;
       const handle = backend_gfx.nativeTextureHandle(backend_id);

   would fail to compile with a `BackendTextureId` vs `BackendTextureId`
   mismatch. Same diamond that bit `backend_input`
   (labelle-assembler#271), reachable through `gfx` only now that the
   texture ids are nominal types.

   This one is invisible to every check in this repo: it only appears when
   the assembler splices the fragment into a generated build.

Verified the probe bites: restoring the old sentinel fails with
"expected type 'backend_contract.BackendTextureId', found 'comptime_int'".
All four CI steps green locally.

* fix(probe): instantiate the video player with a stub, not the real decoder

The previous probe used `VideoPlayer(DesktopVideoDecoder)`, which pulled
`video/desktop.zig` into analysis — and that ACTIVATED its
ffmpeg-dependent tests, failing the macOS runner on
`decodeAudioPcm: ... FfmpegEncodeFailed`. main is green, so the probe
caused it; the type change never did.

Incidentally that confirms the collection rule this PR has been fighting:
a file's inline tests become live exactly when the file is genuinely
analysed. Referencing the alias was not enough; instantiating the
generic was.

A minimal stub decoder compiles the same `Player.init` body without
dragging in a decoder's test surface. It omits `decodeFramePlanes` on
purpose — `Player` gates the GPU plane path on `@hasDecl`, so leaving it
out keeps the stub to the smallest contract that still exercises `init`.

Verified the probe still bites: restoring the old `u32` sentinel fails
with "expected type 'backend_contract.BackendTextureId', found
'comptime_int'". All four CI steps green locally.

+++ src/material_golden.zig
@@ -293,7 +293,7 @@ fn renderScene() void {
             origin,
             0,
             gfx.white,
-            .{ .effect = .palette_swap, .uniforms = .{ .aux_texture = lut.id, .aux_count = 4 } },
+            .{ .effect = .palette_swap, .uniforms = .{ .aux_texture = lut.id.toInt(), .aux_count = 4 } },
         );
 
         // Col 3: dissolve — the solid sprite burned away by the built-in
